chore: Migrate talent synth.py to bazel (#103)
As suggested in https://github.com/googleapis/java-talent/pull/100, this PR only updates the `synth.py` itself to use bazel. 

The updated generated code will be published separately by Yoshi, because it includes breaking changes in resource names, caused not by the bazel migration but by the changes in the generator itself.